+++ src/modules/Dashboard/Container/DashboardContainer.js
@@ -1,8 +1,6 @@
 import { Space, Tag } from "antd";
 import React from "react";
 import { connect } from "react-redux";
-import * as AdminActions from "../../Admin/Store/AdminActions";
-import * as RoleActions from "../../Roles/Store/RolesActions";
 import { store } from "../../../app/ConfigureStore";
 import CButtonAntd from "../../../components/CButton/CButtonAntd";
 import { setGlobalLoading } from "../../App/Store/ComponentAction";
@@ -28,10 +26,10 @@ const DashboardContainer = (props) => {
   }
 
   React.useEffect(() => {
+    initializeApp();
     store.dispatch(setGlobalLoading(false));
     ListServiceActions.getTopTenService();
     getDashboardProgressData();
-    initializeApp();
   }, []);
 
   if (Object.keys(dashboardData).length > 0) {