+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Changelog v.2.4.4
+- fix: Check role when allowed filter branch
+
 ### Changelog v.2.4.3
 - fix: Hide filter cabang on menu customer by roles
 - fix: Takeout 'Rejected' button when status transaction is 'Approved'